fix: rename entity-schema storage to object-literal storage
